+++ web/src/common/constants/fileFormats.js
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export const pcFileFormats = [".e57", ".las", ".laz"];
+export const pcFileFormats = [".e57", ".las", ".laz", ".ply"];
 export const cadFileFormats = [
     ".step",
     ".dwg",
@@ -22,7 +22,6 @@ export const modelFileFormats = [
     ".glb",
     ".stl",
     ".3ds",
-    ".ply",
     ".fbx",
     ".dae",
     ".wrl",
@@ -54,7 +53,7 @@ export const onlineViewer3DFileFormats = [
     ".stl",
     ".obj",
     ".off",
-    ".ply",
+    //".ply" //- Excluded as it will be shown on the point clouder viewer instead
     ".wrl",
     //".fcstd", //- Excluded by default due to license restrictive sub-library use. Enable if you accept this license. Install the appropriate library listed in documentation.
     //".ifc", //- Excluded by default due to license restrictive sub-library use. Enable if you accept this license. Install the appropriate library listed in documentation.